pyproject.toml
[build-system]
requires = ['setuptools>=42']
build-backend = 'setuptools.build_meta'

[project]
name = "swebench"
dynamic = ["version"]
authors = [
    {name = "SWE-bench team", email = "support@swebench.com"},
]
description = "The official SWE-bench package - a benchmark for evaluating LMs on software engineering"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["nlp", "benchmark", "code"]
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3 :: Only",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "beautifulsoup4",
    "chardet",
    "datasets",
    "docker",
    "ghapi",
    "GitPython",
    "modal",
    "pre-commit",
    "python-dotenv",
    "requests",
    "rich",
    "tenacity",
    "tqdm",
    "unidiff",
    "fastcore",
    "fire",
]

[project.optional-dependencies]
datasets = [
    "protobuf",
    "sentencepiece",
    "tiktoken",
    "transformers",
    "openai",
    "anthropic",
    "jedi",
]
inference = [
    "torch",
    "anthropic",
    "jedi",
    "openai",
    "peft",
    "protobuf",
    "sentencepiece",
    "tiktoken",
    "transformers",
    "triton",
    "flash_attn",
    "requests",
]
test = [
    "pytest",
    "pytest-cov",
]
docs = [
    "mkdocs",
    "mkdocs-material",
    "mkdocs-glightbox",
    "mkdocstrings",
    "mkdocstrings-python",
    "mike",
    "pymdown-extensions",
    "mkdocs-include-markdown-plugin",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.dynamic]
version = {attr = "swebench.__version__"}

[tool.setuptools.packages.find]
where = ["."]
namespaces = false

[tool.uv.workspace]
members = [
    "launch",
]

[project.urls]
"Documentation" = "https://github.com/swe-bench/SWE-bench"
"Bug Reports" = "http://github.com/swe-bench/SWE-bench/issues"
"Source Code" = "http://github.com/swe-bench/SWE-bench"
"Website" = "https://swebench.com"
